Fix Error Migrate

+++ database/migrations/2025_10_02_024821_add_is_active_to_users_table.php
@@ -8,16 +8,21 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            // Tambahkan kolom boolean is_active, default 1 (true)
-            $table->boolean('is_active')->default(true)->after('password');
-        });
+        // Hindari error kolom duplikat jika sudah ada
+        if (!Schema::hasColumn('users', 'is_active')) {
+            Schema::table('users', function (Blueprint $table) {
+                // Tambahkan kolom boolean is_active, default 1 (true)
+                $table->boolean('is_active')->default(true);
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('is_active');
-        });
+        if (Schema::hasColumn('users', 'is_active')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('is_active');
+            });
+        }
     }
 };
